pyproject.toml
[project]
name = "my-langchain-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.0.1",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-openai>=1.1.0",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
]
